Create framework and testing thereof
Framework is what used to be the scripts around solving/converting
questions. Now tested separately for coverage as well.
