+++ .coveragerc
@@ -10,7 +10,7 @@ exclude_lines =
     # re-enable default
     pragma: no cover
 
-    # silly input in problems (say)
+    # silly input
     raise ValueError
 
     # script-code
