+++ project_euler/framework/convert
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+python -m project_euler.framework.convert "$@"
