+++ project_euler/framework/test_convert.py
