+++ project_euler/solutions/problems.py
@@ -9,6 +9,7 @@ for filename in os.listdir(os.path.split(__file__)[0]):
 
 numbers = sorted(numbers)
 slow_numbers = {
+    '1': 0.03
     # due to Travis being slower than most machines.
     # will give warnings if more time used, but won't fail
 }
