--- project_euler/solutions/test_solve.py
@@ -1,12 +1,0 @@
-import pytest
-
-from .solve import solve, SolveException
-
-
-def test_solve_valid_problem():
-    solve(2)
-
-
-def test_solve_invalid_problem():
-    with pytest.raises(SolveException):
-        solve('Here be dragons...')
