+++ setup.cfg
@@ -25,5 +25,5 @@ packages =
     project_euler.library
     project_euler.solutions
 scripts =
-    project_euler/solutions/solve.py
-    project_euler/solutions/convert.py
+    project_euler/framework/solve
+    project_euler/framework/convert
