+++ tox.ini
@@ -11,5 +11,6 @@ deps =
     flake8
 commands =
     pytest --cov-config .coveragerc --cov=project_euler/library --cov-report html project_euler/library
-    pytest -n 3 -r w --no-cov --durations=10 project_euler/solutions
+    pytest --cov-config .coveragerc --cov=project_euler/framework --cov-report html project_euler/framework
+    pytest -n 3 -r w --no-cov --durations=10 project_euler/solutions/test_solutions.py
     flake8
